Update Dashboard total

+++ src/utils/helpers.tsx
@@ -1,4 +1,4 @@
 export const formatDate = (date: string) => {
     let myDate = new Date(date)
-    return myDate.toLocaleDateString('en-GB', {year: 'numeric', month: 'long', day: 'numeric'});
+    return myDate.toLocaleDateString('en-GB', {year: 'numeric', month: 'short', day: 'numeric'});
 }
